✨ Made neovim-config a lot nicer
This moved the configuration of **all** plugins into a single file
and added lazy loading to most of them. This also makes configuring and
their keybinds a lot nicer.

Debuggers, formatters and linters can now also be installed and used
from within Mason and hotkey usage is now better explained

Signed-off-by: MaFeLP <60669873+MaFeLP@users.noreply.github.com>